SonarCube suggested renames
