+++ client/src/App.js
@@ -2,7 +2,7 @@ import React, {Fragment} from 'react';
 import "./assets/css/argon-dashboard-react.min.css"
 import './styles/App.css';
 import Header from './components/header';
-import Routes from "./components/routing";
+import Routes from "./components/Routes";
 
 const App = () => (
     <div className="App">
