+++ client/src/components/StoredTradesTable.js
