+++ client/src/components/pages/NewTradePage.js
